3D print files for simple box
.STL and .3MF files added to 3D print a simple box just large enough to house the PicoW and relay. The lid is designed to friction fit.

These were prepared in Bambu Studio and printed on a Bambu A1, but should easily print on any printer.

Printed in PLA
0.4mm Nozzle
0.12mm Quality
25% Sparse Infill
No Support

Once printed, the PicoW and the relay were attached to the box using a dab of hot glue to secure them in place. A self tapping screw (or other mounting method) can then be used to mount the box to the location you have chosen.